Combined navigation reducers to one containing an object
