+++ App/package.json
@@ -74,6 +74,7 @@
     "@types/yup": "^0.29.13",
     "cypress": "^7.7.0",
     "nodemon": "^2.0.12",
+    "redux-devtools-extension": "^2.13.9",
     "ts-node": "^10.1.0",
     "webpack": "^5.44.0"
   }
