--- App/src/reducers/navCollapsed.ts
@@ -1,23 +1,0 @@
-// Function definitions - Defining what parameters and the object structure
-export const setNavCollapsed = (collapsed: boolean = true) => {
-	return {
-		type: 'TOGGLE',
-		payload: collapsed,
-	};
-};
-
-// Defining the type of the action in order for the reducer to know the content of the action
-type ActionType = ReturnType<typeof setNavCollapsed>;
-
-// Defining the reducer, which contains the functionality for each of the functions defined above
-// using action.type to identify the function.
-const navCollapsedReduser = (state: boolean = false, action: ActionType) => {
-	switch (action.type) {
-		case 'TOGGLE':
-			return !state;
-		default:
-			return state;
-	}
-};
-
-export default navCollapsedReduser;
